refactor(GroupInfoService) : Group 정보 조회 서비스 패키지 이동
- Group 정보 조회 관련 클래스를 'org.noostak.group.application.info' 패키지로 이동
- 관련 import 정리 및 패키지 경로 수정
